fix(web): move the delete handler out of the route-imported effects module
The composed loader made the route import account.effects statically, so
the whole module rode in the client graph — and its deleteAccountHandler
kept its dynamic account-binding import alive as an emitted chunk,
dragging plugin-call, the auth runtime and the email templates (css-tree
included) into dist/client. The client-boundary assert rightly refused
the build.

The split the assert asks for: account.effects.ts now holds only the
pure reads the loader calls, and deleteAccountHandler — the session gate
plus the plugin binding, the two things that must never ship — lives in
account-delete.ts behind the server fn's dynamic import.
